+++ CHANGELOG.md
@@ -43,6 +43,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated STM32H7 series yaml to support newly released chips. (#1011)
 - Debugger: Removed the CLI mode, in favour of `probe-rs-cli` which has richer functionality. (#1041)
 - Renamed `Probe::speed` to `Probe::speed_khz`.
+- `ram_download` example now uses clap syntax.
 
 ### Fixed
 
@@ -60,6 +61,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Support the new `defmt` 0.3 `DEFMT_LOG` environment variable.
   - Requires `probe-rs/vscode` [PR #27](https://github.com/probe-rs/vscode/pull/27)
   - Debugger: Improved RTT reliability between debug adapter and VSCode (#1035)
+  - Fixed missing `derive` feature for examples using `clap`.
 
 ## [0.12.0]